+++ server/app.py
@@ -4,9 +4,10 @@ from flask import Flask, jsonify
 from flask_cors import CORS
 
 from models import db, connect_db
-from main import scrape_mynintendo, message_discord, delete_old_records, get_items, get_changes, load_items, get_item_images
+from main import scrape_mynintendo, message_discord, get_items, get_changes, load_items, get_item_images
 from errors import CustomError, CSSTagSelectorError
 from routes.check_api import check_api
+from routes.main_api import main_api
 
 import dotenv
 dotenv.load_dotenv()
@@ -25,6 +26,7 @@ app.config['DEBUG_TB_INTERCEPT_REDIRECTS'] = False
 
 connect_db(app)
 
+app.register_blueprint(main_api, url_prefix="/api")
 app.register_blueprint(check_api, url_prefix="/api")
 
 @app.get('/')
@@ -45,63 +47,6 @@ def show_home_page():
         "last_change": last_change
     }
     return jsonify(display)
-
-
-@app.get('/api/scrape')
-def call_scrape_fn():
-    raw_item_elements = load_items()
-    items = get_items(raw_item_elements)
-    results = scrape_mynintendo(items)
-
-    if results["items"] != "No changes.":
-        message_discord(results["items"])
-
-    return jsonify(results["timestamp"])
-
-
-@app.get('/api/delete')
-def delete_records():
-
-    results = delete_old_records()
-    return f"Deleted {results} entries."
-
-
-@app.get("/api/get-items")
-def call_get_items():
-    raw_item_elements = load_items()
-    items = get_items(raw_item_elements)
-
-    return jsonify(items)
-
-
-# @app.get("/api/check-fly")
-# def check_fly():
-#     try:
-#         return jsonify("API is running.")
-#     except Exception as err:
-#         print(err)
-#         raise CustomError("Fly is down.")
-
-
-# @app.get("/api/check-scraping")
-# def check_scraping():
-#     try:
-#         raw_item_elements = load_items()
-#         get_items(raw_item_elements)
-#         return jsonify("API Scraping is running.")
-#     except Exception as err:
-#         print(err)
-#         raise CustomError("Scraping function failed.")
-
-
-# @app.get("/api/check-db")
-# def check_fly_db():
-#     try:
-#         get_changes()
-#         return jsonify("API DB is running.")
-#     except Exception as err:
-#         print(err)
-#         raise CustomError("DB function failed.")
 
 
 @app.errorhandler(404)
